fix: disable preview media in picture item ai
